verkiezingen_bot/requirements.txt
# Scraper
requests
beautifulsoup4

# PDF parsing
pymupdf

# Embeddings (lokaal)
sentence-transformers

# Vector store
faiss-cpu

# LLM
openai
python-dotenv

# UI
streamlit

# Feedback (Google Sheets)
gspread
google-auth

# Utilities
tqdm
